// src/components/Quizzes/interfaces/materia.interface.ts
// Generated by https://quicktype.io

export interface IMateria {
    idMateria:        string;
    tipo:             string;
    idCiclo:          string;
    idProfesor:       string;
    idGrupo:          string;
    descripcionCiclo: string;
    descripcion:      string;
    __typename:       string;
}